CSRF som middleware-fil
