+++ src/backend/backend/auth/session.ts
@@ -1,9 +1,8 @@
 import type { ISessionKonfigurasjon } from '../typer';
-import type { Express, NextFunction, Request, Response } from 'express';
+import type { Express, Request, Response } from 'express';
 
 import { RedisStore } from 'connect-redis';
 import cookieParser from 'cookie-parser';
-import crypto from 'crypto';
 import session from 'express-session';
 import redis from 'redis';
 
@@ -13,6 +12,7 @@ import {
     hentErforbindelsenTilValkeyTilgjengelig,
     settErforbindelsenTilValkeyTilgjengelig,
 } from '../utils';
+import { csrfBeskyttelse, fjernUtgåtteTokens, genererCsrfToken } from './middleware';
 
 const redisClientForAiven = (sessionKonfigurasjon: ISessionKonfigurasjon) => {
     const pingHvertFjerdeMinutt = 1000 * 60 * 4; // Connection blir ugyldig etter fem minutter, pinger derfor hvert fjerde minutt
@@ -34,42 +34,6 @@ const redisClientForAiven = (sessionKonfigurasjon: ISessionKonfigurasjon) => {
         pingInterval: pingHvertFjerdeMinutt,
     });
     return redisClient;
-};
-
-type CsrfTokenData = {
-    token: string;
-    opprettet: number;
-};
-
-const csrfTokens = new Map<string, CsrfTokenData>();
-const MAKS_TOKEN_ALDER = 24 * 60 * 60 * 1000; // 1 dag
-
-const genererCsrfToken = (sessionId: string) => {
-    const token = crypto.randomUUID();
-    csrfTokens.set(sessionId, { token, opprettet: Date.now() });
-    return token;
-};
-
-const verifiserCsrfToken = (sessionId: string, mottattToken: string): boolean => {
-    const tokenData = csrfTokens.get(sessionId);
-    if (!tokenData) return false;
-
-    const tokenAlder = Date.now() - tokenData.opprettet;
-    if (tokenAlder > MAKS_TOKEN_ALDER) {
-        csrfTokens.delete(sessionId);
-        return false;
-    }
-
-    return tokenData.token === mottattToken;
-};
-
-const fjernUtgåtteTokens = () => {
-    for (const [sessionId, tokenData] of csrfTokens.entries()) {
-        const tokenAlder = Date.now() - tokenData.opprettet;
-        if (tokenAlder > MAKS_TOKEN_ALDER) {
-            csrfTokens.delete(sessionId);
-        }
-    }
 };
 
 // Fjern utgåtte tokens hver time
@@ -147,29 +111,5 @@ export default (app: Express, sessionKonfigurasjon: ISessionKonfigurasjon) => {
         return res.json({ csrfToken });
     });
 
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    app.use((req: Request, res: Response, next: NextFunction): any => {
-        const ikkeSikreMetoder = ['GET', 'HEAD', 'OPTIONS'];
-        if (ikkeSikreMetoder.includes(req.method)) {
-            return next();
-        }
-
-        const sessionId = req.session.id;
-        const csrfToken = req.headers['x-csrf-token'];
-
-        if (!sessionId) {
-            return res.status(401).json({ error: 'Ingen aktiv sesjon' });
-        }
-
-        if (!csrfToken || typeof csrfToken !== 'string') {
-            return res.status(403).json({ error: 'CSRF-token mangler' });
-        }
-
-        if (!verifiserCsrfToken(sessionId, csrfToken)) {
-            logError(`Ugyldig CSRF-token for sesjon ${sessionId}... IP= ${req.ip}`);
-            return res.status(403).json({ error: 'Ugyldig CSRF-token' });
-        }
-
-        next();
-    });
+    app.use(csrfBeskyttelse);
 };
